Separate tools and utils
